// src/modules/users/types/profile.types.ts
export enum AdminStatus {
  ACTIVE = 'Active',
  SUSPENDED = 'Suspended',
  INACTIVE = 'Inactive',
}

export enum Role {
  ADMIN = 'Admin',
  SUPER_ADMIN = 'Super Admin',
}

export enum UserType {
  PATIENT = 'Patient',
  SPECIALIST = 'Specialist',
}

export enum Gender {
  MALE = 'Male',
  FEMALE = 'Female',
}

export enum VerificationStatus {
  VERIFIED = 'Verified',
  UNVERIFIED = 'Unverified',
  SUSPENDED = 'Suspended',
}